web/requirements.txt
Cython==0.29
Django==2.1.2
django-rest-framework==0.1.0
djangorestframework==3.9.0
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.3.0
pytoshop==1.2.0
pytz==2018.5
six==1.11.0
